+++ src/debug.cpp
@@ -10,6 +10,11 @@ $on_mod(Loaded) {
             devtools::label("Icon Kit Utilities\n");
             devtools::label(fmt::format("Current mode: {}", node->m_colorMode).c_str());
         });
+        devtools::registerNode<ItemInfoPopup>([](ItemInfoPopup* node) {
+            devtools::label("Icon Kit Utilities\n");
+            devtools::label(fmt::format("Current type: {}", ICON_NAMES[static_cast<int>(node->m_unlockType)]).c_str());
+            devtools::label(fmt::format("Current ID: {}", node->m_itemID).c_str());
+        });
     });
 }